// src/app/models/car.ts
export interface Car {
    _id: string; 
    ref: string;
    slug: string; 
    name: string;
    body_type: string;
    engine: string;
    mileage: string;
    fuel_type: string;
    transmission: string;
    door_count: string;
    image_car: string;
    price: number;
    description: string;
    desc_excerpt: string;
    date_online: string;
    date_offline: string;
    currency: string;
    contact_phone: string;
    contact_email: string;
}